refactor: merge singular nio/channel into plural nio/channels
