+++ src/commonMain/kotlin/borg/trikeshed/userspace/UserspaceWrappers.kt
@@ -1,6 +1,6 @@
 package borg.trikeshed.userspace
 
-import borg.trikeshed.userspace.nio.channel.Channel
+import borg.trikeshed.userspace.nio.channels.UringChannel
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
@@ -10,7 +10,7 @@ import kotlinx.coroutines.launch
  * Coroutine integration around the userspace Channel.
  */
 class ChannelRunner(
-    val channel: Channel,
+    val channel: UringChannel,
     val scope: CoroutineScope,
 ) {
     private val pendingOps = mutableMapOf<Long, CompletableDeferred<SelectionResult>>()
