+++ src/commonMain/kotlin/borg/trikeshed/userspace/nio/channels/UringChannels.kt
@@ -1,4 +1,4 @@
-package borg.trikeshed.userspace.nio.channel
+package borg.trikeshed.userspace.nio.channels
 
 import borg.trikeshed.userspace.ChannelsImpl
 import borg.trikeshed.userspace.FileImpl
@@ -9,9 +9,9 @@ import borg.trikeshed.userspace.FunctionalUringFacade
 /**
  * Channel factory — backed by expect/actual [ChannelsImpl].
  */
-object Channels {
-    fun open(entries: Int = 256): Channel =
-        Channel(FunctionalUringFacade(entries, openUserspaceChannelBackend(entries)))
+object UringChannels {
+    fun open(entries: Int = 256): UringChannel =
+        UringChannel(FunctionalUringFacade(entries, openUserspaceChannelBackend(entries)))
 
     fun socket(domain: Int, type: Int, protocol: Int): File =
         File(ChannelsImpl.socket(domain, type, protocol))
