+++ src/commonMain/kotlin/borg/trikeshed/userspace/nio/channels/spi/ChannelOperations.kt
@@ -4,7 +4,7 @@ import borg.trikeshed.userspace.nio.ByteBuffer
 import kotlin.coroutines.CoroutineContext
 
 /**
- * Platform channel/socket factory — replaces [borg.trikeshed.userspace.nio.channel.Channels] + ChannelImpl expect.
+ * Platform channel/socket factory — replaces [borg.trikeshed.userspace.nio.channels.UringChannels] + ChannelImpl expect.
  *
  * io_uring submission/completion ring abstraction: prepare operations, submit batch, wait for completions.
  */
